Added pics handling to instructor list users page efficiency improvements to unenroll
git-svn-id: http://imathas.googlecode.com/svn/trunk@405 c89b4f0b-ac2a-0410-9773-c9071ee4f95d

+++ includes/unenroll.php
@@ -19,6 +19,8 @@ function unenrollstu($cid,$tounenroll,$delforum=false,$deloffline=false,$unwithd
 			$threads[] = $rw2[0];
 		}
 	}
+	$threadlist = implode(',',$threads);
+	$forumlist = implode(',',$forums);
 	
 	$assesses = array();
 	$query = "SELECT id FROM imas_assessments WHERE courseid='$cid'";
@@ -26,6 +28,7 @@ function unenrollstu($cid,$tounenroll,$delforum=false,$deloffline=false,$unwithd
 	while ($row = mysql_fetch_row($result)) {
 		$assesses[] = $row[0];
 	}
+	$aidlist =  implode(',',$assesses);
 	
 	if (count($tounenroll)>0) {
 		$curdir = rtrim(dirname(__FILE__), '/\\');
@@ -36,6 +39,37 @@ function unenrollstu($cid,$tounenroll,$delforum=false,$deloffline=false,$unwithd
 		while ($row = mysql_fetch_row($result)) {
 			$gbitems[] = $row[0];
 		}
+		$gblist = implode(',',$gbitems);
+		//new
+		$stulist = "'".implode("','",$tounenroll)."'";
+		if (count($assesses)>0) {
+			$query = "DELETE FROM imas_assessment_sessions WHERE assessmentid IN ($aidlist) AND userid IN ($stulist)";
+			mysql_query($query) or die("Query failed : $query" . mysql_error());
+			$query = "DELETE FROM imas_exceptions WHERE assessmentid IN ($aidlist) AND userid IN ($stulist)";
+			mysql_query($query) or die("Query failed : $query" . mysql_error());
+			deleteasidfilesbyquery(array('assessmentid'=>$assesses, 'userid'=>$tounenroll));
+		}
+		if (count($gbitems)>0) {
+			$query = "DELETE FROM imas_grades WHERE gbitemid IN ($gblist) AND userid IN ($stulist)";
+			mysql_query($query) or die("Query failed : $query" . mysql_error());
+		}
+		if (count($threads)>0) {
+			$query = "DELETE FROM imas_forum_views WHERE threadid IN ($threadlist)  AND userid IN ($stulist)";
+			mysql_query($query) or die("Query failed : $query" . mysql_error());
+		}
+		
+		$query = "DELETE FROM imas_students WHERE userid IN ($stulist) AND courseid='$cid'";
+		mysql_query($query) or die("Query failed : $query" . mysql_error());
+		/*
+		foreach ($tounenroll as $uid) {
+			foreach ($assesses as $aid) {
+				deleteasidfilesbyquery(array('assessmentid'=>$aid, 'userid'=>$uid));
+			}
+		}*/
+				
+		
+		//old, replaced with above
+		/*
 		foreach ($tounenroll as $uid) {
 			$query = "DELETE FROM imas_students WHERE userid='$uid' AND courseid='$cid'";
 			mysql_query($query) or die("Query failed : " . mysql_error());
@@ -57,27 +91,36 @@ function unenrollstu($cid,$tounenroll,$delforum=false,$deloffline=false,$unwithd
 				mysql_query($query) or die("Query failed : " . mysql_error());
 			}
 		}
+		*/
 		
 		
 	}
-	if ($delforum) {
+	if ($delforum && count($forums)>0) {
+		$query = "DELETE imas_forum_threads FROM imas_forum_posts JOIN imas_forum_threads ON imas_forum_posts.threadid=imas_forum_threads.id AND imas_forum_posts.posttype=0 WHERE imas_forum_threads.forumid IN ($forumlist)";
+		mysql_query($query) or die("Query failed : " . mysql_error());
+			
+		$query = "DELETE FROM imas_forum_posts WHERE forumid IN ($forumlist) AND posttype=0";
+		mysql_query($query) or die("Query failed : " . mysql_error());	
+		/* //old
 		foreach ($forums as $fid) {
 			$query = "DELETE imas_forum_threads FROM imas_forum_posts JOIN imas_forum_threads ON imas_forum_posts.threadid=imas_forum_threads.id AND imas_forum_posts.posttype=0 WHERE imas_forum_threads.forumid='$fid'";
 			mysql_query($query) or die("Query failed : " . mysql_error());
 			
 			$query = "DELETE FROM imas_forum_posts WHERE forumid='$fid' AND posttype=0";
 			mysql_query($query) or die("Query failed : " . mysql_error());	
-		}
+		}*/
 	}
 	if ($deloffline) {
-		$query = "DELETE from imas_gbitems WHERE courseid='$cid'";
+		$query = "DELETE FROM imas_gbitems WHERE courseid='$cid'";
 		mysql_query($query) or die("Query failed : " . mysql_error());
 	}
-	if ($unwithdraw) {
-		foreach ($assesses as $aid) {
+	if ($unwithdraw && count($assesses)>0) {
+		$query = "UPDATE imas_questions SET withdrawn=0 WHERE assessmentid IN ($aidlist)";
+		mysql_query($query) or die("Query failed : " . mysql_error());
+		/*foreach ($assesses as $aid) {
 			$query = "UPDATE imas_questions SET withdrawn=0 WHERE assessmentid='$aid'";
 			mysql_query($query) or die("Query failed : " . mysql_error());
-		}
+		}*/
 	}
 		 
 }
